BF: rename Routine was leading to broken experiment loading fixes #1248
The name wasn't being changed in the saved file. The problem was that (for
old compatibility reasons) the name is stored in two places for a Routine
and only one of these was being changed. Routine.name is now a property
that automatically updates Routine.params['name'] when it gets set (but
the reverse is not true)