ENH: New sound option sounddevice
sounddevice is another portaudio option but pure python, supporting py3 and better maintained than pysoundcard.

This implementation sets up an appropriate stream as soon as the sound is created but plays zeros until the sound is played (whereas in the other backends so far we've been creating the stream at the moment we play() ). This seems to have better timing than the existing solutions (although some of them could use the same approach if we reworked them).

Not yet implemented:
  - sound from array (so movie3 won't work using this yet)
  - loops etc
There are also likely to be some bugs, but the future is bright!