NF: Allow delayed displaying of DlgFromDict
DlgFromDict dialogs would always be shown
automatically on instantiation. This is problematic
for testing. This commit adds a new `show` keyword
that can be used to disable this behavior. The dialog
can be displayed at a later time by explicitly calling a
newly added show() method. The default behavior remains
unchanged.