ENH: added psyexpCompile as a script (1st step in supporting useVersion)
For builder to support useVersion properly we need to be able to
write the script in the correct as well as *running* the script in the
right version.

To do this we need to spawn a new process and compiler the script from
builder.experiment.Experiment there (or we would have to do lots of
reload commands on our existing instance).

The new psyexpComile script allows this. Specify your in/out filenames
and optionally a version for useVersion and allow it to switch before
the script compile stage.