add seeding capability to psychojs random loops
This adds a line to the default index.html template for web
experiments. See pull request with same title in the psychojs
repository for changes to the psychojs code to implement this.

This uses seedrandom() from https://github.com/davidbau/seedrandom
and will require adding the seedrandom.min.js from that repository
to the psychojs/js/vendors/ folder in the psychojs.zip file
that is packaged with the psychopy standalone app.