ENH: ImageStim now shows nothing if image=None rather than a color
This seems more intuitive and makes it easier to have no-image
trials (by leaving the value empty in the conditions file).
Previously that gave a white square, so you had to set opacity=0
as well.

I've added the option to set the value of image
to be "color" which will achieve the same as setting it to None
previously did (ie you'd get a texture that is ones rather than
no texture at all)