RF: numpy is complaining about equality tests again
image values can be numpy arrays, which can't be compared with None
or we get a complaint that this is ambiguous but now we can't even
do that using `value in (None, "None", "none")` so now we have to have

`if value is None or value == "None" or value == "none":`

to avoid the warning. Annoying!