NF: Add support for == and != comparisons
This commit implements the __eq__() and __ne__ (== and !=) methods
for `DataHandler` and `_BaseTrialHandler`.

Example:

In [1]: from psychopy.data import TrialHandler
   ...:
   ...: t1 = TrialHandler([dict(foo=1)], 2)
   ...: t2 = TrialHandler([dict(foo=1)], 2)
   ...:
   ...: print(t1 == t2)
   ...:
True

Specializations were required for

- TrialHandler2: ignore RNG object
- QuestHandler and PsiHandler, as they use contrib classes (see below)

QuestHandler and PsiHandler rely on classes implemented in contrib
modules; these should be left unchanged. Because an == or != comparison
on two QuestHandler / PsiHandler instances will eventually descent into
their accompanying QuestObject / PsiObject, __eq__() and __ne__() have
to be implemented for these objects as well. However, I did not want to
modify the contrib code, so I first tried an approach with
monkey-patching, but this was ugly and not very intuitive. I now
subclass QuestObject (as QuestObject_) and PsiObject (as PsiObject_),
respectively, and add the required __eq__() and __ne__() methods there.
Accordingly, QuestHandler and PsiHandler will now instantiate
QuestObject_ and PsiObject_, respectively.

Why all this? Mostly because it will ease writing tests and debugging.
For example, when saving a TrialHandler object to disk and later reading
it back in, it can now be compared to an existing in-memory object very
easily.