NF: Add JSON serialization for StairHandler
StairHandler and derived objects can now be serialized as JSON.
Serialization of TrialHandler* is currently not supported, as
json-tricks cannot properly handle the required DataHandler, because
it is derived from a dict. See [1] for details.

This is a first step in allowing us and our users to move away from
pickle for long-term storage.

References:
[1] https://github.com/mverleg/pyjson_tricks/issues/30