Update qtgui.py
`isinstance` is used to check if `thisType` is an instance of `basestring` in this case, `thisType` is a class itself, not an instance of a class. If `thisType` is `<type 'unicode'>`, then using `isinstance(thisType, basestring) returns False- which appears to not be the intended function. However using `issubclass(thisType, basestring) returns True because `unicode` is a subclass (not an instance) of basestring.