RF/BF: Use OrderedDict in importConditions
We used to use an ordinary dict to store the condition parameters, which
could lead to non-deterministic behavior. Specifically the order of the
keys would be system-dependent, and it differed between Py2 and 3. This
in turn lead to inconsistent column orders in the data files.

(cherry picked from commit e97cb5c)