FF: fixed dialog validator message
There were clashes where checks on the name and checks on the syntax of
code would overwrite each other. Needed a dict in the dialog (not in the
validator) to store warnings according to the name of the field.

Then the message presents the first element in this dict of warnings