ENH: Ensure we're using pythonw when running on macOs with Anaconda
On macOS Anaconda installations, GUI apps have to be started using
`pythonw` instead of `python`. Since distutils sets the shebang of
`psychopyApp.py` to a `python` executable, users cannot simply invoke
`psychopyApp.py` to start the GUI (will result in an error).

A potential workaround could have been to try to detect whether
`psychopyApp.py` was run using `pythonw` or `python`. In the former
case, no intervention would be required; in the latter, we would spawn
a new `pythonw` interpreter.

However, it is impossible to determine whether `pythonw` or `python`
was used to run the script (`sys.executable` returns `python` in either
case). But if we invoked `pythonw` ourselves, we could set an
environment variable indicating that indeed the current Python
process is of the `pythonw` type.

So, my solution is to

- check if we're running in an Anaconda interpreter on macOS
  - check if a `PYTHONW` environment variable is set to `True`
    - is yes, start the GUI
    - if no, run `psychopyApp.py` in `pythonw` with the environment
      variable `PYTHONW` set to 'True'

So on macOS with Anaconda, `psychopyApp.py` is invoked two times: first
by the user calling the script, and then it "runs itself again" using
`pythonw`. This may not the most elegant approach, but appears to be
working properly.

Closes #1326.