BF: movie3 seek behavior + audio
Bug fixes for moviestim3 covering issues with seeking behavior as well as adding functionality for sounddevice. In particular, added checks to take advantage of sounddevice's expanded functionality, and safeguards to prevent seek from seeking invalid time-stamps (which was leading to crashes) and doubling-up audio. As a bonus, it is now possible to seek while paused, which previously would mess up the audio.