BF: Movie3 try/except when finding next frame
As a further upgrade to the resilience in my last bug fix, this version adds a try/except statement to the updateFrame function such that if it tries to find a frame and fails, rather than crashing outright it just looks for the next frame and raises a warning. If there's a way to break this, I haven't found it, and I've recreated previous crash conditions several times.