WIP/RF: moving Experiment/Components/Params... to be outside the app
This is going to be important so that we can load psyexp files and
compile scripts without the app having to load (which means we can use
useVersion() properly with the Builder)

Unfortunately this doesn't solve the problem yet. There's another issue
that all component params are translated using wx.locale at the point of
creation rather than at the point of use. Need to discuss with @hsogo
how hard this will be to switch (so that params are translated when used
to create the Dlgs)