FF: Supporting gamma on Py3 and Py2
On py2 we used to do 0xFFFFFFFF to force the return value of
pyglet Windows device context (_dc) into the right structure
but that didn't work on Py3 - it wouldn't convert the types
correctly. Tried another solution using codecs to convert from
hex, but that's no good either. So now we have orig solution on
Py2 and new conversion on Py3