BF: ShapeStim was changing vertices when doing stim.vertices+=val
This one was interesting!

The tesselate functions add further vertices (presumably caused by
the need to loop back?) but further tesselate functions should then
act on the _original_ requested vertices not the tesselated ones.

Solution: keep a separate (hidden) variable for tesselated verts and
this is only updated after the call to stim.vertices += val