BF: rare crash caused by frame advancing on loadMovie
https://discourse.psychopy.org/t/rare-and-strange-conflict-between-sound-file-and-movie-file-with-ffmpeg/3906/6
Short version: _updateFrameTexture() used to advance by one frame unless the status of the object was "paused", but because _updateFrameTexture() is called on loadMovie, it was impossible to call "pause" without it auto-advancing one frame. Calling "play" then resets the timestamp to 0, causing ffmpeg_reader to re-initialize the whole movie, which sometimes caused memory issues, lag, and crashes.

The fix is that _updateFrameTexture() now only advances if the status is PLAYING. This should be safe thanks to my previous fixes (lines 315-322), which add a layer of insurance if it goes looking for a frame with a negative timestamp or one that otherwise does not exist. Based on my testing this seems much more stable and possibly even better-performing.