NF: Adds new 'Variable' component to Builder.
The variable component allows users to define variables in builder. A variable can be defined and updated at the beginning of the experiment, routine, or on every frame. Values are stored in a dict, and if requested the

values are written to the data file on every trial using trial handler. The value of the variable can be saved as either a constant, a variable from every routine, as a list of every value from each frame, or the final

routine or frame value for each trial. Documentation included in pull request.