BF: Removes unnamed columns during importConditions.
The importConditions function uses Pandas to import data from csv or excel files. However, if a column header is empty, an incorrect Value Error is raised telling the user that column headers cannot contain punctuation or spaces. This is because Pandas names unnamed columns on import with names that contain spaces. To fix, unnamed columns are removed from the dataframe, assuming that columns are unnamed because the user does not want to use that column in the experiment.