FIX: Builder Static Component across routines
When compiling an experiment from the builder, if a component that executes in a routine (routine B) updates during the static component from a previous routine (routine A) the compilation fails. This results from the experiment pointing to the updating component as if it was in the same routine as the static component, which isn't always going to be true. With this edit, the `updatesList` in the static component will know the correct routine that the updating component resides and can be properly compiled from builder to script.