BF: Fixed hex colors with Window and stimuli together
Commit #1696 made the handling of hex colors correct for stimuli, but
broke it for windows. #1822 did the opposite.

The real fix is not to subtract the [1,1,1] but then make clear to the
Window that the values are rgb255 (i.e. not zero-centered) unlike rgb
colors.

This provides that fix and adds tests for hex colors in stimuli and
Window to the test suite to prevent regression