FIX: sounddevice stereo auto detection
sound device backend was not running the code to determine whether audio was mono or stereo. This is because the `if/elif` block of code to do this was indented too far in and never able to run. Moving this block of code out one indentation level allows the stereo attribute to be detected and properly reassigned based on the sound array's features.