BF: This commit bypasses parallel Python init if version is None.
The fix refactors the psyexpCompile script into a function which can be run
either from command line or as an import. This speeds up the script compiler
when no version is selected. The imported function cannot be
used to set the version of PsychoPy. Version can only be set from
command line.


Signed-off-by: dvbridges <david-bridges@hotmail.co.uk>