Extended class "DlgFromDict" with the following features:
(1) An "order" parameter (either bool or list) can be used to define the order of the displayed fields. This allows for a user defined order of input fields in the dialog.

(2) A "labels" parameter can be used to define labels which are displayed instead of the keys. This allows for comprehensible captions for input fields in the dialog while maintaining brief key strings.

(3) A "preserve_type" parameter can be used to enforce preservation of the variable types provided in the dictionary, if at all possible. This remedies the following problem, currently occurring under Windows. Say, the dictionary is:

  dictionary = {'Parameter1', 99}   # the 99 is an int

The dialog will display an input field labeled "Parameter1" with a pre-entered value of 99. If the user then deletes the latter "9" and inputs "7", the returned dictionary will be:

  dictionary = {'Parameter1', 97}   # the 97 is still an int

However, if the user deletes the "99" - thus having the field be empty a some point during user input - and then re-enters "97", the dictionary will be:

  dictionary = {'Parameter1', '97'}   # the 97 has become a string

So, in the former case, the provided value will remain an int. In the latter case, since an empty input field cannot be converted to int, the variable type will switch to string and remain so.