ENH: tyring different form of lazy import for pavlovia_ui.sync and git
Problem, when we import git (or pavlovia_ui.sync, which has a git
subclass) even if we use try...except it triggers a mac to complain that
it needs git to be installed (whether or not the user will actually
try to sync a project).

Our existing lazy_import doesn't import pavlovia_ui.sync properly
for some reason.

This one does and seems very neat so trying it out for git and sync