BF: setRedirectUrl function should only exist when URLs are provided.
The setRedirectURL function takes string arguments from online experiment
settings, however the default setting was redirecting to nonexisting URLs.
This fix changes the default value in the URL params, and only writes
the setRedirectURL function in the JS if an entry is provided for one of
the URL params in exp. settings.