ENH: Return stdout, stderr of pythonw invocation during app startup
Running `psychopyApp.py` used to silently fail (return code 0, no
error message) for me on macOS with a Miniconda installation. Turned
out that one dependency (`wxpython`) was missing, but I never got
to see the error message because the output of the `pythonw` executable
that we invoke via `shellCall()` was silently discarded. This commit
changes this behavior: we will now retrieve `stdout` and `stderr`, and
"re-print" them to the appropriate streams.