NF: Add Experiment.requireImport() method
This is a more generic way to add import statements to a Builder-
generated experiment. Previously, only `requirePsychopyLibs()` was
available, which would only allow for `from psychopy import xxx`;
whereas the new `requireImport()` can import `from` and `as`
arbitrary packages/modules/names.

`requirePsychopyLibs()` was modified such that it now acts as a
convenience methods for `requireImport()`, and _could_
theoretically be removed some time in the future.