RF/TESTS: Changes py2js conversion of params and adds py2js tests
Numbers in lists with unary operators are converted to tuples, inconsistent with numbers without unary operators existing as lists. Now, elements of list/tuple with unary operators are converted to nested list, unless a binary operator exists.
Also, some tuples were not converted to lists, and now py2js checks for tuple in params that are lists, and converts to list (which was not happening previously).
py2js tests are added to check conversions of parameters.