ENH: Add `Experiment.runOnce()` method
This method allows to add arbitrary code to the beginning of the
experiment script (directly below the imports) which will only
be executed a single time. This is for example useful if a Component
requires a certain initialization routine, e.g. activation of hardware,
and this procedure shall be executed only once even if the
Component appears multiple times in the experimental flow.

cf. https://github.com/psychopy/psychopy/pull/2141#issuecomment-447295089