PKG: Use entry_points to create PsychoPy executable
This automatically creates an executable with the right interpreter,
in the right place, on all platforms. This is also going to help us
provide `conda` packages, as `conda-build` also supports a similar
directive, even for platform-independent (`noarch`) packages, and
ensures the script is run via `python.app` on macOS if desired.

https://setuptools.readthedocs.io/en/latest/setuptools.html#automatic-script-creation
https://conda.io/docs/user-guide/tasks/build-packages/define-metadata.html#python-entry-points