PKG, TESTS: Vendor `configobj` `master`
The latest release of `configobj` is already a bit old and does not
support Python 3.7. For that reason, we switched to using the `master`
branch – at least for testing – a couple months back. Now when working
on the "bring PsychoPy to conda-forge" project, it turned out that
this is in fact a quite problematic approach as `conda-forge` does not
accept "non-released" software packages (meaning we would not be able
to package `configobj` `master` for the Python 3.7 build).

I discussed the situation with
the `conda-forge` folks and they suggested the easiest approach would
probably be to "vend(or)" `configobj` `master`, i.e., ship it with
PsychoPy, at least until a new version with the fixes we need has been
published.

This is exactly what this commit does. Note that it also retains
support for local `configobj` installations (which take precedence
over the contrib package); yet, I (temporarily) dropped `configobj`
from the package requirements.

cc @kastman