BF/RF: Fixes sync problem occuring when user was not actively logged in.
If a project was opened, but the user had not logged in, and tried to sync,
the user was asked to create a project, even though the project existed.
The failed sync then resulted in the user being logged in and the sync
working on the next attempt. This fix attempts to log the user in on a sync
if the user is known, or in the case of multiple users where the namespace
is not indicative of the user account, uses the project name to determine
membership to known user and which known user should be logged in. Also
increases number of items listed to 100 when retrieving user projects from
Pavlovia. Fixes and refactors warning messages.

# Conflicts:
#	psychopy/projects/pavlovia.py