BF: Fixes how EOL are converted in XML Builder file
Previous problems with EOLS were found when legacy Mac EOLs were used in
code components, because PsychoPy could not convert those EOLs into the
correct XML format. This fix now also checks for legacy Mac EOLS in code
components only, but in doing so, covers three EOL styles (\\r, \\n and \\r\\n for legacy Mac, Unix
and Windows). The previous fix for controlling and converting EOLs using
preferences may still be useful, because it creates uniform EOLs in the
PsychoPy when it is compiled from Builder.

The prefs for newline convention were causing a global reconfig of EOLS
in components that use CodeBox class e.g., text and RatingScale.
This fix also keeps the EOLs as standard UNIX in components that
use CodeBox, but are not code components. This way, there are no
differences presenting text locally or online that may be caused by
newline prefs.