BF: Code component was not using Paste method from base code class
This fix adds a keycheck when using paste shortcut and redirects
to the codeEditorBase paste method. It also enables the
context menus for both Coder and Code Component to be directed
to the Paste function in the codeEditorBase.

The context menu does not work when switching tabs in Code Component.
The context menu history is reset on switch - so in practice you cannot
paste etc between tabs when using the context menu only (works for keys).
This needs fixing. In the mean time, the keyboard shortcut is independent
of the context menu fixes, and can be cherry picked if the menu is not
wanted.