BF: Allow .xlsm files to be read using `importConditions`
This fix allows .xlsx files to be loaded as conditions files in PsychoPy.
Older versions of openpyxl allowed cell contents to be accessed using the
coordinates of a cell e.g., 'A1`, however from 2.0, cell contents are
accessed using row, col coordinates with a starting index of 1. This fix
also checks for openpyxl versions < 2.0, to determine how to read cells.
Fixes #2309