NF: Allow routines to be disabled
Note that because of the  way PsychoPy adds routines to the flow,
there is currently no easy way to allow users to disable individual
representations of the same routine in the flow. If a routine is
placed in the flow multiple times, disabling one of them will disable
_all_ occurrences.

More background info on this:
When adding a routine to the flow, PsychoPy simply adds this routine
to a list. When adding another representation of the same routine to
the flow, effectively the very same object will get added again -- they
both point to the same object in memory. Modifying one will
automatically alter the other too (since, well, it is not "another",
but "the same" routine). To fix this, we'd have to come up with a
solution to store metadata for each representation of a routine in the
flow individually.