BF: No longer assume user account to log-in
Following #2258, this fix now raises a user log-in dialog to make user
log-in if they are not logged in but want to sync or run a study from
PsychoPy. Previously in #2258, we attempted to make this non-intrusive
by finding usernames associated with projects, and make the log-in attempt
for them. This can cause problems when multiple user accounts exist
that contain forks/clones of the same projects - see discourse issue
https://discourse.psychopy.org/t/6955