BF: Setting mouse position on retina displays
numpy array of self.win.size returns raw pixels, so double the pixels that the mouse position is working in (dammit Apple). As a result, while dividing by two gets you the center of a non-retina display, you have to divide by four to get the center of the actual window on a retina display if you want to set the mouse position with any accuracy. Also you have to divide the pixel number you get from units2pix by 2 if you want it to do THAT part accurately. Basically if they're using a retina display, divide the whole newPosPix calculation by 2 and you get something that behaves the way you expect when you call mouse.setPos, even if it's ugly under the hood.