BF: Removes unnecessary imports from Cedrus/ioLabs safely
This fixes a problem where cedrus and ioLabs imports would wipe keyboard
imports if they appeared in the same experiment. This fix adds new
functionality to the experiment class by giving ability to check more
abstractly whether a particular type of component exists in an experiment.
For example, whether a component of type keyboard exists. With this new
functionality, Cedrus and ioLab boxes will only delete keyboard imports
if there are no keyboard components in the experiment.