BF: Fixes how escape is used to quit the experiment online (#2359)
* BF: Fix escape key behaviour for online experiments

Escape to quit behaviour was hardcoded into the routine code for JS. This
fix removed hardcoded escaping using "escape" and is now determined by
experiment settings option.

* BF: Fixes how escape is used to quit the experiment online

Previously, the JS code would always quit the experiment if escape was
pushed. This fix allows user to control whether or not escape quits the
experiment, using the experiment settings.

- Tidies up some spacing inconsistencies between conditionals in JS
outputs

- Only writes code if code exists - this prevents blank lines from being
written into the code

* BF: Check escape membership with keyListStr, not KeyList

Also refactors image stim code for consistency, and resets correct tests
after removing excess blank lines

* BF: Keep escape key code in keyboard code, even if escape not in keylist

We want to keep this code because even when escape is not in keyList, it
may still be a valid key when all keys are allowed i.e., when keyList
is None.

* BF: Add colorSpace back as a parameter in Python image code.

This param was lost during the refactor.