BF: Fix keyboard use of basestring or str depending on Python version
When a variable is used as a keylist and is not a list, tuple or np array,
it is evaluated to determine whether it is a string. In Python 2,
basestring is the comparison type, but basestring does not exist in
Python 3. This fix changes the comparison type depending on Python version
running.