BF: better cleanup of application on sys.exit
The AUI manager doesn't clean itself up properly on exit (we need to
call Uninit() on the AUI managers). That is already done by the
CoderFrame and BuilderFrame closeFrame() methods BUT if the app wasn't
explicitly closed (closed by sys.exit or end of script) then it didn't
get called.

Now using atexit to make sure the application quit() function is called
even if closed by sys.exit (or end of script on travis)