BF: Create videostim objects on each routine for looping variables
JS MovieStim will not initialise without a video, and so creating a movie
object with a variable name replaced with an undefined type, when variables
are used, will not work. This fix follows the Python code and creates the
movie object on in experimentInit if constant is used, else creates the
movie object at beginning of each routine if vars are used.