BF: Ensure events cleared and clock reset for KB when using visual Sync
When KB was synced to visual, the KB clock was not reset and events
not cleared before first response was collected, only when
a response happened immediately after keyboard status was changed
to "STARTED". The clearEvents call now added to callOnFlip when visual
sync for KB is True. The new waitOnFlip var does not allow
getKeys to be called if the initial callOnFlip, used for resetting clock
 and clearing events, has not been completed.