BF: Stops expression2js running on text component text containing parens
This fixes an error that causes a syntax error when JS is compiled
and a text component has text containing parenthesis. As this is text,
the round brackets do not need converting to a list. If code is used
in the text field, the correct py2js conversion is performed by the
Param class.