BF: Save data when calling loop.finished in nested loop
In a nested loop, with the outerloop isTrials is False, and the
inner loop isTrials is True, a participant ending their inner loop
on the first iteration, but has repeating inner loop n times based on number
of trials in the outerloop, will not have data saved. This is because
breaking the inner loop on the first iteration does not call nextExtry(),
and the outer loop where isTrials is False does not call nextEntry().
Calling at the end of the task will only write the last trial recorded.
This fix calls nextEntry in endLoopIteration() when currentLoop
 is true and orphan data exists.