BF: Leave out Variable components when tracking start/stop times
This fixes an issue where Variable components cause AttributeErrors
when datatypes, that do not support attribute assigment, are used in the
component. The issue has arisen following new changes that the track start and
stop times of PsychoPy components. This fix stops Variable comps from being
added to the list of routine components that need start/stop times
recording. Fixes #2499.