BF: Call endLoopIteration when isTrials is False so loop can be broken
The endLoopIteration was not being called when isTrials was false, meaning
that the loop could not be broken using loop.finished. Adding the
endLoopIteration function to loops where isTrials is false means that the
a new arg `isTrials` must be passed, because `isTrials` never actually
exists in the `thisTrials` arg, so can not satisfy the else if statement,
and thisTrial can be both defined and undefined when isTrials is false,
meaning nextEntry could be called when it should not be. Now,
endLoopIteration will only call nextEntry if isTrials is true if not
trying to break loop manually.