BF: long-standing issue where lots of print statements crash builder run
The issue was that the wx.Process wasn't polling the stdout and its
buffer got full. We've moved to use subprocess.popen instead which doesn't
seem to have the same limits on output buffer.

Currently the output is then only read at the end of the run, but that's
OK for most Builder studies?