RF: new PTB pref audioLatency changed to audioLatencyMode
It might be confused with an actual latency value (e.g. millisecs)
rather than the mode/class as intended.

Also converted to use number only as setting, and then translate-able
labels are added during presentation of the relevant dialogs (so as not
to interfere with the usage behind the scenes).