ENH: Provide visual feedback on sync button for success or failure
This enhancement will provide visual feedback about whether or not a sync
was successful. Users see a green or red globe for success or fail, which
resets after a set time. This is useful now, but also in preparation
for cleaning up the dialogs. In future, we can pass the sync status
information to either the alerts system, or write the information into
the sync panel of the project details dialog. A red globe will indicate the
user to check the alert log or project details page.