ENH: Adds code component syntax check using alerts package
Now, code component syntax is tested for Python and JS each time a script
is compiled. Any errors are allowed to pass but alerts are formatted
with informative information, such as which tab contains the error.
The information is sent to the stdoutFrame and saved in the alerts
log folder. Also adds string formatting values to the size, position and
time checks,  checks time values can be tested correctly.