ENH: Make errorHandler app property and change error stream capture
For now, we only need to capture the error stream for the error handler.
The output of its prints are directed to the stdout, so for seeing warnings
it only matters where that is pointing to. In future, it may be good to
send outputs and errors to the runner frame only, one source for all
information.