BF: Fixes JS snapshot code, and declares vars in nested functions
One issue, if you use a routine in a loop and therefore have the nested
functions for when the function is scheduled with the snapshot, and then use
that same routine outside of a loop, the nested function in the routine
function does not get called (usually called by the loop scheduler),
so the experiment will not progress in this situation. A fix requires
that the flowScheduler calls the nested function if it is not in a loop.