BF: Fixes alerttool error from refactor.
The refactor did not catch use of variables in all cases. Variables cause
errors when converting to float, since variable names are strings.
This fix checks each value can be converted to a float
(meaning it is a number), but is now more specific so start and end
times are checked independently. Also handles syntax errors in dialog
boxes, so real time syntax errors are not raised and printed to console,
but caught and thus allowing the dialog to perform its syntax warnings.