BF: uses dlg to warn about metapensiero, rather than using code comments
This fix will only set code type to "Auto->JS" when code box is created
if metapensiero is installed, else if falls back onto "Both". A dialog is
used to warn if the user tries to use "Auto->JS", and then their choice
is reversed when the dialog is closed. This stops existing method, where
comments are added to the users code.