Update to runner cwd
See issue #2186. The new runner changed how the cwd was set so that it no longer used the file's directory as the working directory, and instead defaulted to the app folder. This fix uses the fileName string and uses rfind to locate the last instance of a folder delimiter ('\\' for windows, '/' for everything else) to set the working directory to the path of the current file. Tested on Mac, not yet tested in Windows.