FF: Code Components need to know about the new themes
Previously the removal of prefs['codeFont'] meant that the code boxes
in Component dialogs, and also Code Components themselves, crashed.

That's fixed here by providing the app as a variable and using
app._codeFont
to provide some valid monospaced font.

On code boxes within standard dialogs I think that's actually fine as
a solution.

For Code Components though (which are built on BaseCodeEditor) we end up
with an ugly standard monospace font instead of a nice one (consolas on
PC, Monaco on mac, Ubuntu Mono or DejaVu Sans Mono on Linux)