DOC: Fixed link to TrialHandler, added link to TrialHandlerExt
The link to TrialHandler class was not working, but StairHandler was, therefore I changed to the same format.

In the past years I have seen multiple researchers asking for an oddball trial handler. I developed a solution quite some time ago and it still seems to work, so I added the link in the docs.