Update eyetracker.py
Added support for pupil diameter in mm.
As stated in official API field LPUPILD and RPUPILD (which can be enabled by SEND_EYE_LEFT and SEND_EYE_RIGHT at init) contains pupil diameters in meters (sic!). These measures need to be converted and it is done by multiplying by 1000 (duh).
This code was tested with real hardware and it works.