BF: harmless but annoying error on shutdown caused by garbage collect order
This appears to be something to do with pyglet event module and the order
of garbage collection.

Creating any of the stimuli in this pull request and then ending the script
will yield:
```
ModuleNotFoundError: import of pyglet halted; None in sys.modules
```

It's harmless because the script is closing anyway but annoying to see.

Fixes #1816 improves on #2085