FF: Make pyo parsing of devices more robust
The methods to get input/output sound devices info in the pyo backend
rely on a parsing of the output of `pyo.pa_list_devices()`. It splits
the fields (name, host api index, latency, default sr) by splitting
on commas. This is fragile as the name can contain commas (e.g.
"Logitech USB Headset: Audio (hw:1,0)"), which leads to `IndexError:
list index out of range` error.

`pyo` provides a method `pa_get_devices_infos()` [0] that returns 2
dictionaries for input and output devices, almost exactly in the
expected format as returned by the current implementation of
`_query_devices()`. The only differences are in the fields values: no
units ("Hz", "s"), do not split the name to exclude the content after a
colon and the delay has many digits (vs 6 in the initial
implementation).

This commit switches to using `pa_get_devices_infos()` with a small
adaptation of the format to keep the same output (but the name) as the
previous implementation was doing.

/!\ Behavior change:

The name of a sound device is now more precise, e.g.

Before: "Logitech USB Headset"
After: "Logitech USB Headset: Audio (hw:1,0)"

It makes the identifier more unique, as it contains the hardware position,
for example if we have several identical headsets connected.

[0] https://github.com/belangeo/pyo/blob/274ed61d6664d44d23416957e781b8afbb559589/src/engine/pyomodule.c#L139

Fixes #3365